[PWM] Fixed getDutyCycle function

+++ src/shared/drivers/timer/PWM.cpp
@@ -84,8 +84,8 @@ void PWM::setDutyCycle(TimerUtils::Channel channel, float dutyCycle)
 
 float PWM::getDutyCycle(TimerUtils::Channel channel)
 {
-    return static_cast<float>(timer.readCaptureCompareRegister(channel) /
-                              timer.readAutoReloadRegister());
+    return static_cast<float>(timer.readCaptureCompareRegister(channel)) /
+           static_cast<float>(timer.readAutoReloadRegister());
 }
 
 void PWM::setTimerConfiguration()
